Added GetPorts utility method and it's unit test.
GetPorts retrieves the starting and load balancer port numbers from environment variables.
If environment variables are not set or contain invalid values, it defaults to predefined ports.
STARTING_PORT is the first port for n number of server instances, and LOAD_BALANCER_PORT is for the load balancer.
Defaults: startingPort :8000 and loadBalancerPort :8080