X-Forwarded-For:
We set the X-Forwarded-For header to r.RemoteAddr, which contains the IP address of the client making the original request. This helps the target server identify the true origin of the request.
X-Forwarded-Proto:

We check if r.TLS is nil. If not, the connection is using TLS/HTTPS, so we set the X-Forwarded-Proto header to "https". Otherwise, we set it to "http". This lets the target server know whether the original request was secure.